Add Code::Blocks project file for the code from Lecture 4
This is for Linux only; project files for C::B on Windows and
Visual Studio coming soon.

You need to copy this file as 'fray.cbp' in the root directory of
the project, then open it from there. Please note that 'fray.cbp'
is intentionally git-ignored, because it is not a good practice to
store "directly-openable" project files in a source-control
system (if it weren't like this, and fray.cbp was in git, if you
opened it with a different version of C::B, it may automatically
add/remove things to the projfile without you actually doing
anything, producing unnecessary noise and wories).

I.e. GIT is only committed to tracking what humans do. SkyNet
looks after the machines :)

More info on how projfiles are used/setup and how you can get
a happily compiling environment on your machine will be put
in the course site.