Add VisualStudio vcxproj file and a minor build fix
Similar to the C::B project file for Linux, the VS project file has to
be copied in the root directory of the project and opened from there.
The project builds and copies the correct SDL shared library to the
target output directory of the build, making manual copy redundant.

Removed SDL header included in main.cpp, since it was no longer needed
there and thus avoiding another conditional compilation.

Conditionally include the SDL header from different location on Windows,
because the headers are distributed in flat directory, thus fixing the
include only with managing include paths becomes error-prone.