Make the sphere with a glassy shader, using a layered shader.
The top layer is a Reflection shader, using opacity texture
(which is FresnelTexture).
The bottom layer is a Refraction shader.

A slight problem with this very versatile approach of defining
shaders is that you may create combinations that don't make sense.
For example, this glass shader needs to have a IOR parameter,
which is used in both the Refraction shader and the FresnelTexture.
You must use the same value in both places, but nothing enforces
you to do so.